Improve Rules of Attribution for Private Works
Previously required that all Private Works may only be attributed to God without explicit attribution, and that the creator of the Private Work cannot place any name on the Work for the purposes of attribution. Such a rule is more problematic than not, especially in the case of leaked Private Works, as no one would know who it came from; and in the case of a breach of license, no one could be held accountable.

Instead, the rule has been changed to better reflect the Catholic position of this company, as it is also in a certain sense unjust to improperly attribute or omit due attribution for a Work, so Private Works now require explicit attribution.

Sub-section d was added to Article III as well. This is the primary intent of attributions in this legal framework, and the benefits are obvious.